TronECKey GetPublicAddressPrefix()返回值类型报错，进行修改调试，暂时不知道是否正确

+++ src/TronNet/TronECKey.cs
@@ -48,7 +48,7 @@ namespace TronNet
 
         internal byte GetPublicAddressPrefix()
         {
-            return _network == TronNetwork.MainNet ? 0x41 : 0xa0;
+            return (byte)(_network == TronNetwork.MainNet ? 0x41 : 0xa0);
         }
 
         /// <summary>
